// frontend/src/client/models/UserProfileOut.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserProfileOut = {
    id: number;
    full_name: string;
    username: string;
    email: string;
    description: (string | null);
    profile_image: (string | null);
};